Make class-php-ico a thin wrapper around a true PSR-2 compliant class
